feat(13): adds links for basic and graphic
